chore: Update dependencies (#869)
* Update dependencies

This includes updating Tokio to 1.x.
The feature flags of Tokio have been changed and `dns` was merged into `net`.

`fast-socks5` was used in two different versions because of a dependency in async-smtp. This gets addressed in https://github.com/async-email/async-smtp/pull/36, but in the meantime we can divert to the fix from the fork at https://github.com/amaurymartiny/async-smtp/tree/am-fast-socks.

* Use SmallRng, which is Send+Sync

Hyper requires shared types to be Send+Sync. The threaded rng is not.
Replacing it with SmallRng does the trick,
which is satisfactory for creating random mail addresses.

* cargo fmt

* Mark tests with tokio_test

* Update hyper

* Update test suite to Tokio 1.x